// src/api/ambulance-management/models/DispatchPriority.ts
// @ts-nocheck
/* tslint:disable */
/* eslint-disable */
/**
 * Ambulance Management API
 * API contract for ambulance management modules including fleet and dispatch management.
 *
 * The version of the OpenAPI document: 1.0.0
 * Contact: your_email@stuba.sk
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * Triage priority assigned to the dispatch.
 * @export
 */
export const DispatchPriority = {
    Low: 'LOW',
    Medium: 'MEDIUM',
    High: 'HIGH',
    Critical: 'CRITICAL'
} as const;
export type DispatchPriority = typeof DispatchPriority[keyof typeof DispatchPriority];


export function DispatchPriorityFromJSON(json: any): DispatchPriority {
    return DispatchPriorityFromJSONTyped(json, false);
}

export function DispatchPriorityFromJSONTyped(json: any, ignoreDiscriminator: boolean): DispatchPriority {
    return json as DispatchPriority;
}

export function DispatchPriorityToJSON(value?: DispatchPriority | null): any {
    return value as any;
}